requirements.txt
numpy==1.19.4
Pillow==8.0.1
streamlit==0.73.1
torch==1.7.1
torchvision==0.2.2.post3
